+++ packages/charts/src/lib/chartContainer/ChartContainer.svelte
@@ -117,7 +117,6 @@
 			<ExportBtns
 				{chartToCapture}
 				{filename}
-				idToPad="captureElement"
 				dataForDownload={data}
 				{dataDownloadButton}
 				{imageDownloadButton}