+++ packages/ui/src/lib/radioButton/RadioButtonGroup.svelte
@@ -53,11 +53,11 @@
 	export let buttonsHidden = false;
 </script>
 
-<div class="flex flex-col space-y-0.25">
+<div class="flex flex-col space-y-0.5">
 	{#if !buttonsHidden}
 		<Button variant="text" class="!px-0" on:click={() => (selectedId = '')}>Clear</Button>
 	{/if}
-	<div class={orientation === 'vertical' ? 'flex flex-col space-y-0.25' : 'flex space-x-3'}>
+	<div class={orientation === 'vertical' ? 'flex flex-col space-y-1' : 'flex space-x-3'}>
 		{#each options as option}
 			<RadioButton
 				id={option.id}